refactor(lambda-promtail): move SNS and SQS event processing to s3.go
